+++ shpc/main/container/singularity.py
@@ -23,6 +23,7 @@ class SingularityContainer(ContainerTechnology):
 
     # The module technology adds extensions here
     templatefile = "singularity"
+    command = "singularity"
 
     # Singularity container features
     features = {
@@ -71,7 +72,7 @@ class SingularityContainer(ContainerTechnology):
             logger.exit("Found more than one sif in module folder.")
         return sif[0]
 
-    def add(self, sif, module_name, modulefile, template, wrapper_template, **kwargs):
+    def add(self, sif, module_name, modulefile, template, **kwargs):
         """
         Manually add a registry container.
         """
@@ -112,7 +113,6 @@ class SingularityContainer(ContainerTechnology):
             template,
             parsed_name=parsed_name,
             features=kwargs.get("features"),
-            wrapper_template=wrapper_template,
         )
         self.add_environment(module_dir, {}, self.settings.environment_file)
         logger.info("Module %s was created." % (module_name))
@@ -131,7 +131,7 @@ class SingularityContainer(ContainerTechnology):
         config_features=None,
         features=None,
         version=None,
-        wrapper_template=None,
+        config=None,
     ):
         """Install a general container path to a module
 
@@ -170,20 +170,22 @@ class SingularityContainer(ContainerTechnology):
 
         # Option to create wrapper scripts for commands
         module_dir = os.path.dirname(module_path)
-        if self.settings.wrapper_scripts and aliases:
-            self._generate_wrapper_scripts(
-                wrapper_template, 
-                aliases, 
-                module_dir, 
-                features, 
-                container_sif=container_path, 
+
+        # Wrapper scripts can be global (for aliases) or container specific
+        wrapper_scripts = None
+        if self.settings.wrapper_scripts["enabled"] is True:
+            wrapper_scripts = shpc.main.wrappers.generate(
+                aliases=aliases,
+                module_dir=module_dir,
+                features=features,
+                container=self,
+                image=container_path,
+                config=config,
             )
 
         # Make sure to render all values!
         out = template.render(
-            singularity_module=self.settings.singularity_module,
-            singularity_shell=self.settings.singularity_shell,
-            bindpaths=self.settings.bindpaths,
+            settings=self.settings,
             container_sif=container_path,
             description=description,
             aliases=aliases,
@@ -195,12 +197,8 @@ class SingularityContainer(ContainerTechnology):
             deffile=deffile,
             creation_date=datetime.now(),
             name=name,
-            tool=parsed_name.tool,
-            registry=parsed_name.registry,
-            repository=parsed_name.repository,
-            envfile=self.settings.environment_file,
-            wrapper_scripts=self.settings.wrapper_scripts,
-            wrapper_subdir=self.wrapper_subdir,
+            parsed_name=parsed_name,
+            wrapper_scripts=wrapper_scripts,
         )
         shpc.utils.write_file(module_path, out)
 
@@ -354,7 +352,7 @@ class SingularityContainer(ContainerTechnology):
         """
         Given a test file, run it and respond accordingly.
         """
-        command = ["singularity", "exec", image, "/bin/bash", test_script]
+        command = [self.command, "exec", image, "/bin/bash", test_script]
         result = shpc.utils.run_command(command)
 
         # We can't run on incompatible hosts